requirements.txt
beautifulsoup4
bs4
certifi
charset-normalizer
click
Flask
Flask-Cors
gunicorn
idna
itsdangerous
Jinja2
MarkupSafe
requests
six
soupsieve
urllib3
Werkzeug
cryptography
pycryptodome
